requirements.txt
keyboard==0.13.5
openai==1.56.2
Pillow==11.0.0
playsound==1.3.0
PyAutoGUI==0.9.54
python-dotenv==1.0.1
torch==2.4.1
torch_directml==0.2.5.dev240914
